QUASAR: Add 'Test SMTP Settings' button to basic config to test settings
The button will pull an object comprised of the current Domain, Address,
Port, Password and Recipients list and push it through a non-cron version 
of the python mailing script to see if, using the provided settings, we can 
get a mail from the Nova system to the addresses on the recipients list.

Decoupled SMTP_AUTH checkbox from modifying client side elements,
as we now know that people can and will set up SMTP servers on 
non-standard ports.