QUASAR: Fix issues with cron.* calling scripts as root
Now modifying crontab directly, allowing scripts to run as the user as 
opposed to as root. The reason the alerts weren't working before was that
root's os.environ["HOME"] is not where maildaemon.lock is, which would 
cause the novasendmail script to fail early.

Also removed an extension, moved a script, updated the Makefile to clean
the email stuff up (it wasn't before) and modified main.js to reflect the 
change to non-root land.