Fix compatibility bug with newer versions of libpcap
You're not supposed to call pcap_compile before calling pcap_activate.
Older versions of libpcap didn't check for this, but new versions
throw an error if you try to do this.

I split the SetFilter method into SetFilter and ActivateFilter, where
ActivateFilter is only called during a live update of the filter or
when the capture is started, and SetFilter is called alone if the filter
is being set before the capture has been started.