Fix stored XSS vulnerability
There was a stored XSS vuln found in Nova and publically disclosed. The attacker can inject arbitrary values into the telnet username/password, which is stored as classification notes, and exposed to the browser without any sanitation. This means that an attacker can inject carefully crafted Javascript into the telnet honeypot login and use it to take control of the Nova admin UI.

http://seclists.org/fulldisclosure/2014/Nov/77

http://webcache.googleusercontent.com/search?q=cache:r_Zb92FYaZAJ:https://cultofthedyingsun.wordpress.com/2014/11/21/datasoft-nova-anti-reconnaissance-system-13-10-0-stored-xss-via-emulated-honeypot-telnet-instance/+&cd=2&hl=en&ct=clnk&gl=us

This fix sanitizes the data before being exposed in the suspect details, logs, or new information pages.